fixing index and creating post an event form
