+++ app/assets/stylesheets/pages/_index.scss
@@ -1,11 +1,11 @@
 // Import page-specific CSS files here.
 @import "home";
 @import "show";
+@import "new";
 
 .main-container{
   width: 325px;
   margin: 0 auto;
-  margin-bottom: 80px;
 }
 
 .banner{
@@ -65,16 +65,16 @@ h1 {
   color: $primary;
 }
 
-// h2 {
-//   font-size: 28px;
-//   font-weight: bold;
-// }
+h2 {
+  font-size: 28px;
+  font-weight: bold;
+}
 
-// h3 {
-//   font-size: 22px;
-//   font-weight: bold;
-// }
+h3 {
+  font-size: 22px;
+  font-weight: bold;
+}
 
-// p {
-//   font-size: 18px;
-// }
+p {
+  font-size: 18px;
+}
