+++ app/assets/stylesheets/pages/_new.scss
@@ -1,0 +1,3 @@
+.create-event-form{
+  margin-bottom: 80px;
+}
