+++ app/javascript/packs/application.js
@@ -27,25 +27,15 @@ import "bootstrap";
 
 // Internal imports, e.g:
 // import { initSelect2 } from '../components/init_select2';
-const downArrow = document.getElementById("more-categories");
-const categories = document.querySelectorAll(".category-group");
-downArrow.addEventListener('click', (event) => {
-  categories.forEach((category) => {
-    category.classList.toggle("hide");
-  });
-  event.currentTarget.classList.toggle("fa-chevron-up")
-});
-
-document.addEventListener('turbolinks:load', () => {
-  // Call your functions here, e.g:
-  // initSelect2();
-});
 
 import { initMapbox } from '../plugins/init_mapbox';
 import { setUpEventSearchPills } from '../plugins/event_pills';
-
+import { initDownArrow } from '../components/init_down_arrow';
+import { initAutocomplete } from '../plugins/init_autocomplete'
 
 document.addEventListener('turbolinks:load', () => {
   initMapbox();
   setUpEventSearchPills();
+  initDownArrow();
+  initAutocomplete();
 })
