+++ app/javascript/plugins/init_autocomplete.js
@@ -1,0 +1,10 @@
+import places from 'places.js';
+
+const initAutocomplete = () => {
+  const addressInput = document.querySelector('location');
+  if (addressInput) {
+    places({ container: addressInput });
+  }
+};
+
+export { initAutocomplete };
