+++ package.json
@@ -9,6 +9,7 @@
     "bootstrap": "^4.5.3",
     "jquery": "^3.5.1",
     "mapbox-gl": "^2.0.0",
+    "places.js": "^1.19.0",
     "popper.js": "^1.16.1",
     "turbolinks": "^5.2.0"
   },
